Up to middle
Edits in writing in Chapters 3 and 4. Page 116 in the paragraph that starts with "One benefit of..." there is a partial line of text halfway through paragraph before the Figure reference. I tried to delete a potential return in the line but this might just be spacing thing and is ok if not fixable. No changes in 5 found.