Edits based on review of PDF
Found a few small changes. Biggest one is to replace y3 in the QQ-plot demo with a different seed version. Somehow this had become a t-dist without heavy tails. This version tells the story better. Data in simy3 and in folder and online, but no code provided for students in this section.